backend/src/requirements.txt
fastapi==0.51.0
uvicorn==0.11.1
motor==2.2.0
python-jose==3.2.0
passlib==1.7.2
bcrypt==3.2.0
email-validator==1.1.1
python-multipart==0.0.5
aiofiles==0.5.0